% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds.R
\docType{data}
\name{cds}
\alias{cds}
\alias{cds_acc_desc}
\alias{cds_accident}
\alias{cds_event}
\alias{cds_gv}
\alias{cds_oa}
\alias{cds_oi}
\alias{cds_pers_pro}
\alias{cds_typ_acc}
\alias{cds_ve}
\alias{cds_veh_pro}
\alias{cds_vi}
\title{NASS Crashworthiness Data System}
\format{something about being llkalk}
\source{
\url{http://www.nhtsa.gov/Data/National-Automotive-Sampling-System-(NASS)/NASS-Crashworthiness-Data-System}
}
\usage{
cds_acc_desc

cds_accident

cds_event

cds_gv

cds_oa

cds_oi

cds_pers_pro

cds_typ_acc

cds_ve

cds_veh_pro

cds_vi
}
\description{
The CDS blah blah blah
}
\keyword{datasets}

