% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nass.R
\docType{package}
\name{nass}
\alias{nass}
\alias{nass-package}
\title{nass: Data from the National Automotive Sampling System (NASS)}
\description{
Data from both the General Estimates System (GES) and
Crashworthiness Data System (CDS) which are representive random samples of
crashes reported by the National Highway Traffic Safety Administration.
}
\section{CDS}{

The Crashworthiness Data System data sets are all have the prefix \code{cds_}.
See \code{\link{cds}}.  Currently only 2015.
}

\section{GES}{

General Estimates System data isn't in this package yet
}

